# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Land-Management-System API",
      "description": "Land-Management-System API",
      "version": "1.0.0"
    },
    "schemes": ["http", "https"],
    "securityDefinitions":{
      "Bearer":{
       "type":"apiKey",
       "name":"Authorization",
       "in":"header",
       "description":"Enter the token for Authorization"
      }
      },
    "paths": {
...
      "/api/reseacher/login": {
        "post": {
          "summary": "Login",
          "description": "Login",
          "tags": ["Authentication"],
          "parameters":[{
            "name":"body",
            "in":"body",
            "description":"reseacher details",
            "required":"true",
            "schema":{
                "type":"object",
                "properties":{
                    "email":{
                   "type": "string",
                    "required": [true,"Please provide email first"],
                    "unique": "true",
                    "lowercase": "true"
                    },
                   "password":{
                    "type": "string",
                    "required": [true,"Please provide password first"],
                    "minLength": [8,"Must be at least 8 characters"]
                    }
                      
                }
            }
        }],
          "responses": {
            "201": {
              "description": "User Logged In"
            }
          }
        }
      },
      "/api/user/login": {
        "post": {
          "summary": "Login",
          "description": "Login",
          "tags": ["Authentication"],
          "parameters":[{
            "name":"body",
            "in":"body",
            "description":"User details",
            "required":"true",
            "schema":{
                "type":"object",
                "properties":{
                    "email":{
                   "type": "string",
                    "required": [true,"Please provide email first"],
                    "unique": "true",
                    "lowercase": "true"
                    },
                   "password":{
                    "type": "string",
                    "required": [true,"Please provide password first"],
                    "minLength": [8,"Must be at least 8 characters"]
                    }
                      
                }
            }
        }],
          "responses": {
            "201": {
              "description": "User Logged In"
            }
          }
        }
      },
      "/api/landOwner/login": {
        "post": {
          "summary": "Login",
          "description": "Login",
          "tags": ["Authentication"],
          "parameters":[{
            "name":"body",
            "in":"body",
            "description":"landOwner details",
            "required":"true",
            "schema":{
                "type":"object",
                "properties":{
                    "email":{
                   "type": "string",
                    "required": [true,"Please provide email first"],
                    "unique": "true",
                    "lowercase": "true"
                    },
                   "password":{
                    "type": "string",
                    "required": [true,"Please provide password first"]
                    
                    }
                      
                }
            }
        }],
          "responses": {
            "201": {
              "description": "landOwner Logged In"
            }
          }
        }
      },
...
  },
...
  }